Move common code for fs-ag-driver to under /syndicate/ag/fs_driver_common
